+++ Package@swift-5.9.swift
@@ -1,0 +1,26 @@
+// swift-tools-version: 5.9
+
+import PackageDescription
+
+let package = Package(
+  name: "JailbreakDetector",
+  platforms: [
+    .iOS(.v15)
+  ],
+  products: [
+    .library(
+      name: "JailbreakDetector",
+      targets: ["JailbreakDetector"]
+    )
+  ],
+  targets: [
+    .target(
+      name: "JailbreakDetector"
+    ),
+    .testTarget(
+      name: "JailbreakDetectorTests",
+      dependencies: ["JailbreakDetector"]
+    )
+  ],
+  swiftLanguageVersions: [.v5]
+)